root/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alabaster==1.0.0
    # via sphinx
attrs==25.3.0
    # via
    #   hypothesis
    #   jsonschema
    #   referencing
babel==2.17.0
    # via sphinx
beautifulsoup4==4.13.5
    # via -r requirements.in
blake3==1.0.5
    # via -r requirements.in
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6
    # via
    #   griffe
    #   pytest
    #   sphinx
docstring-parser==0.17.0
    # via -r requirements.in
docutils==0.21.2
    # via
    #   -r requirements.in
    #   sphinx
fastjsonschema==2.21.2
    # via nbformat
griffe==1.13.0
    # via -r requirements.in
hypothesis==6.138.13
    # via -r requirements.in
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.1.0
    # via pytest
jedi==0.19.2
    # via -r requirements.in
jinja2==3.1.6
    # via sphinx
jsonschema==4.25.1
    # via nbformat
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter-core==5.8.1
    # via nbformat
jupytext==1.17.3
    # via -r requirements.in
lxml==6.0.1
    # via -r requirements.in
markdown-it-py==4.0.0
    # via
    #   jupytext
    #   mdit-py-plugins
markupsafe==3.0.2
    # via jinja2
mdit-py-plugins==0.5.0
    # via jupytext
mdurl==0.1.2
    # via markdown-it-py
nbformat==5.10.4
    # via
    #   -r requirements.in
    #   jupytext
packaging==25.0
    # via
    #   jupytext
    #   pytest
    #   sphinx
parso==0.8.5
    # via
    #   -r requirements.in
    #   jedi
platformdirs==4.4.0
    # via jupyter-core
pluggy==1.6.0
    # via pytest
psutil==7.0.0
    # via -r requirements.in
pygments==2.19.2
    # via
    #   pytest
    #   sphinx
pytest==8.4.1
    # via -r requirements.in
pywin32==311
    # via jupyter-core
pyyaml==6.0.2
    # via jupytext
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via sphinx
roman-numerals-py==3.1.0
    # via sphinx
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
snowballstemmer==3.0.1
    # via sphinx
sortedcontainers==2.4.0
    # via hypothesis
soupsieve==2.8
    # via beautifulsoup4
sphinx==8.2.3
    # via -r requirements.in
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
traitlets==5.14.3
    # via
    #   jupyter-core
    #   nbformat
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   referencing
urllib3==2.5.0
    # via requests
